shell-pop-split-window: Fix bug
Split the frame's main window instead of the root window to avoid violating
a main window invariant in window.el. See Emacs bug 73627 for details.

This bug was fixed on the master branch (for Emacs 31), however, to support
Emacs < 31, we need to work around the problem here. This also fixes #63.